+++ CHANGELOG.md
@@ -1,5 +1,24 @@
 ## [Releases]
 
+## [0.1.1] - 2022-05-05
+
+### Fixed
+
+- Typo on Stream closing error message. ( Thanks to @ook )
+
+## [0.1.0] - 2022-05-05
+
+- Yanked
+
+## [0.1.0.beta.1] - 2022-03-17
+
+### Added
+
+- `/clusterStatus` endpoint support
+- `/healthcheck` endpoint support
+- `/info` endpoint support
+- `/terminate` endpoint support
+
 ## [0.1.0.beta] - 2022-02-24
 
 - Beta release